This project is intended to demonstrate the function of Cross-Origin Resource Sharing (CORS) in a project between Spring Boot and React.
The last LTS version of NPM/Node and Java 17 are in use for this.

The example is based on this tutorial:
https://www.youtube.com/watch?v=HRwlT_etr60 and its associated
repository: https://github.com/danvega/spring-security-cors-demo/blob/master/src/main/frontend/src/App.vue

There are a few differences:

1) This variant is done with React instead of Vue
2) In the SecurityConfig.java file, there is a correction in the
SecurityFilterChain method, compared to the original:

	@Bean
	SecurityFilterChain securityFilterChain(HttpSecurity http) throws
	Exception {

		http
		.authorizeRequests(auth -> auth
				.anyRequest().authenticated())
		.httpBasic(Customizer.withDefaults())
		.cors();
		return http.build();

	}

I've added http.cors() to the configuration to solve a persistent HTTP
401 error that the original example had, at least in my case and in that
of other people who had commented on the project.

The article for the solution is found here:
https://www.baeldung.com/spring-security-cors-preflight

Any commented content represents various attempts to address that CORS
401 error, in the event someone is interested in what did not solve the
issue.